requirements.txt
asyncio >= 3.4.3
aiohttp >= 3.9.3
beautifulsoup4 >= 4.11.2
pytz >= 2022.7.1
requests >= 2.28.2
redis
# elasticsearch
# Levenshtein
fuzzywuzzy
celery[redis]
fastapi[all] >=  0.65.1
uvicorn >=  0.14.0
certifi >= 2021.5.30
pymongo >= 3.11.4
python-decouple >= 3.8
selenium
webdriver_manager
python-json-logger >= 2.0.7

# OpenTelemetry - Distributed Tracing
opentelemetry-api >= 1.25.0
opentelemetry-sdk >= 1.25.0
opentelemetry-instrumentation-fastapi >= 0.46b0
opentelemetry-instrumentation-aiohttp-client >= 0.46b0
opentelemetry-instrumentation-pymongo >= 0.46b0
opentelemetry-instrumentation-redis >= 0.46b0
opentelemetry-exporter-otlp-proto-grpc >= 1.25.0

# Prometheus Metrics
prometheus-client >= 0.20.0
prometheus-fastapi-instrumentator >= 7.0.0
